AI-Powered_Form_filling_Assistant/backend/requirements.txt
flask
flask-cors
easyocr
pymupdf
opencv-python
pillow
numpy
python-dotenv
openai
langdetect